dogstatsd reportHandlers
